# openapi8-8-22.yaml
openapi: 3.0.2
servers:
  - url: http://localhost:8080
info:
  description: Esta es la API que vamos a crear en clase
  version: "1.0"
  title: Delivery API
...
paths:
  /orders:
    get:
      tags: 
        - orders 
      summary: Listado de orders
      description: Devuelve el listado completo de todas los pedidos registradas. Es posible filtrar por la distancia del pedido
      operationId: getOrders
      parameters:
        - name: distance
          in: query
          description: Dsitancia de que se encuentra entre el punto de entrga y el restaurante
          required: false
          schema:
            type: number
            format: int32
            default: 500
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado la comanda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar la order"
                
    post:
      tags:
        - orders
      summary: Registrar una order
      description: Registra una nueva order en el sistema
      operationId: postOrders
      requestBody:
        description: Los datos de la nueva order
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderIn'
      responses:
        '201':
          description: Se ha registrado la order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                
        '400':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
...
  /rider/{riderId}/orders:
    get:
      tags: 
        - orders
        - riders
      summary: Obtiene las orders de un rider
      description: Obtiene la información de las orders que tiene un rider deterniado segun la distancia
      operationId: getRiderOrders
      parameters:
        - name: riderId
          in: path
          description: Identificador del rider
          required: true
          schema:
            type: number
            format: int64
            example: 23
        - name: distance
          in: query
          description: Longitud de las orders que quiere
          required: false
          schema:
            type: number
            format: int32
            default: 0
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'
        '404':
          description: No se ha encontrado ningun order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el order"
...
  /riders:
    get:
      tags: 
        - riders
      summary: Listado de riders
      description: Devuelve el listado completo de todas los riders registrados. Es posible filtrar por el vehoculo del rider
      operationId: getRiders
      parameters:
        - name: vehicle
          in: query
          description: vehiculo en la que viaja el rider
          required: false
          schema:
            type: string
            default: "moto"
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rider'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado un rider con ese vehiculo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el rider"
    post:
      tags:
        - riders
      summary: Registrar un rider
      description: Registra un nuevo rider
      operationId: postRider
      requestBody:
        description: Los datos de un nuevo rider
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiderIn'
      responses:
        '201':
          description: Se ha registrado el rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '400':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
  /rider/{riderId}:
    get:
      tags: 
        - riders
      summary: Obtiene los datos de un rider
      description: Devuelve los datos de un rider determinado
      operationId: getRider
      parameters:
        - name: riderId
          in: path
          description: Identificador del rider
          required: true
          schema:
            type: number
            format: int64
            example: 1
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado el rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el rider"
    delete:
      tags: 
        - riders
      summary: Da de baja un restaurante
      description: Da de baja un restaurante determinado
      operationId: deleteRider
      parameters:
        - name: riderId
          in: path
          description: Identificador del rider
          required: true
          schema:
            type: number
            format: int64
            example: 23
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '404':
          description: No se ha encontrado el rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el rider"
    put:
      tags: 
        - riders
      summary: Modifica un rider
      description: Modifica los datos de un rider determinado
      operationId: modifyRider
      parameters:
        - name: riderId
          in: path
          description: Identificador del rider que se va a modificar
          required: true
          schema:
            type: number
            format: int64
            example: 23
      requestBody:
        description: Los datos nuevos del rider
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiderIn'
      responses:
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '404':
          description: No se ha encontrado el rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el rider"
                
                
    patch:
      tags: 
        - riders
      summary: Modifica el nombre del rider
      description: Modifica el nombre del rider
      operationId: patchRider
      parameters:
        - name: riderId
          in: path
          description: string que cambia el nombre del rider
          required: true
          schema:
            type: string
            example: "Daniel"
      requestBody:
        description: Los datos nuevos del rider
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiderIn'
      responses:
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
        '404':
          description: No se ha encontrado el rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFound'
              example:
                internalError: 1
                message: "No se ha podido encontrar el rider"
components:
  schemas:
    OrderIn:
      properties:
        price:
          type: number
          format: int64
          example: 20.5
        weight:
          type: number
          format: double
          example: 50.5
        ready:
          type: boolean
          example: false
        time:
          type: integer
          format: int32
          example: 30
        distance:
          type: number
          format: int32
          example : 500
        restaurant:
          type: integer
          format: int32
          example: 2
        user:
          type: integer
          format: int32
          example: 1
        rider:
          type: integer
          format: int32
          example: 3
      type: object
    OrderDto:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 23
        - $ref: '#/components/schemas/OrderIn'
...
    RiderIn:
      properties:
        name:
          type: string
          example: "Jose Luis"
        surname:
          type: string
          example: "Lopez"
        vehicle:
          type: string
          example: "bicicleta"
        maxSpeed:
          type: number
          format: int32
          example: 50
      type: object
    Rider:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int64
              example: 23
        - $ref: '#/components/schemas/RiderIn'       
...
    ErrorResponse:
      properties:
        internalError:
          type: number
          format: int32
        message:
          type: string
      type: object
    ResourceNotFound:
      allOf:
        - type: object
        - $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      allOf:
        - type: object
        - $ref: '#/components/schemas/ErrorResponse'
      example:
        internalError: 999
        message: "Internal Server Error"
    BadRequest:
      allOf:
        - type: object
        - $ref: '#/components/schemas/BadRequest'
      example:
        internalError: 400
        message: "BadRequest"